Conflicts:
	func/domain.sh

+++ func/domain.sh
@@ -372,12 +372,14 @@ del_web_config() {
             old=$IP6
             get_web_config_lines $WEBTPL/$1/$WEB_BACKEND/$2 $conf
             sed -i "$top_line,$bottom_line d" $conf
-
-            web_domain=$(grep DOMAIN $USER_DATA/web.conf |wc -l)
-            if [ "$web_domain" -eq '0' ]; then
-                sed -i "/.*\/$user\/.*$1.conf/d" /etc/$1/conf.d/vesta.conf
-                rm -f $conf
-            fi
+        fi
+    fi
+    
+    web_domain=$(grep DOMAIN $USER_DATA/web.conf |wc -l)
+    if [ "$web_domain" -eq '0' ]; then
+        sed -i "/.*\/$user\/.*$1.conf/d" /etc/$1/conf.d/vesta.conf
+        if [ -f "$conf" ]; then
+            rm -f $conf
         fi
     fi
 }
